`message_count` to `u32` in `coconut`, tweaks for `utils`

+++ coconut/src/proof/messages_pok/mod.rs
@@ -318,7 +318,7 @@ mod tests {
             MessagesPoKError::MessageInputError(MessageUnpackingError::MessageIndexIsOutOfBounds(
                 IndexIsOutOfBounds {
                     index: 10,
-                    length: message_count
+                    length: message_count as usize
                 }
             ))
         );
